refactor user router by creating a controller for it
